Rename engine components files
