+++ runner/src/server/plugins/engine/components/index.ts
@@ -1,30 +1,30 @@
 // TODO:- resolution is buggered here. don't use this yet.
 // TODO:- Ideally @model/types map EXACTLY to @engine/components. at the moment they're being required by using str.toLowerCase().
 
-export { default as AutocompleteField } from "./autocompletefield";
-export { default as CheckboxesField } from "./checkboxesfield";
-export { default as Component } from "./component";
-export { default as ComponentCollection } from "./componentcollection";
-export { default as ConditionalFormComponent } from "./conditionalformcomponent";
-export { default as DateField } from "./datefield";
-export { default as DatePartsField } from "./datepartsfield";
-export { default as DateTimeField } from "./datetimefield";
-export { default as DateTimePartsField } from "./datetimepartsfield";
-export { default as Details } from "./details";
-export { default as EmailAddressField } from "./emailaddressfield";
-export { default as FileUploadField } from "./fileuploadfield";
-export { default as Flashcard } from "./flashcard";
-export { default as FormComponent } from "./formcomponent";
-export { default as html } from "./html";
-export { default as InsetText } from "./insettext";
-export { default as List } from "./list";
-export { default as MultilineTextField } from "./multilinetextfield";
-export { default as NumberField } from "./numberfield";
-export { default as Para } from "./para";
-export { default as RadiosField } from "./radiosfield";
-export { default as SelectField } from "./selectfield";
-export { default as TelephoneNumberField } from "./telephonenumberfield";
-export { default as TextField } from "./textfield";
-export { default as TimeField } from "./timefield";
-export { default as UkAddressField } from "./ukaddressfield";
-export { default as YesMoField } from "./yesnofield";
+export { default as AutocompleteField } from "./AutocompleteField";
+export { default as CheckboxesField } from "./CheckboxesField";
+export { default as Component } from "./Component";
+export { default as ComponentCollection } from "./ComponentCollection";
+export { default as ConditionalFormComponent } from "./ConditionalFormComponent";
+export { default as DateField } from "./DateField";
+export { default as DatePartsField } from "./DatePartsField";
+export { default as DateTimeField } from "./DateTimeField";
+export { default as DateTimePartsField } from "./DateTimePartsField";
+export { default as Details } from "./Details";
+export { default as EmailAddressField } from "./EmailAddressField";
+export { default as FileUploadField } from "./FileUploadField";
+export { default as Flashcard } from "./Flashcard";
+export { default as FormComponent } from "./FormComponent";
+export { default as html } from "./Html";
+export { default as InsetText } from "./InsetText";
+export { default as List } from "./List";
+export { default as MultilineTextField } from "./MultilineTextField";
+export { default as NumberField } from "./NumberField";
+export { default as Para } from "./Para";
+export { default as RadiosField } from "./RadiosField";
+export { default as SelectField } from "./SelectField";
+export { default as TelephoneNumberField } from "./TelephoneNumberField";
+export { default as TextField } from "./TextField";
+export { default as TimeField } from "./TimeField";
+export { default as UkAddressField } from "./UkAddressField";
+export { default as YesMoField } from "./YesNoField";
